requirements.txt
# Web frameworks
fastapi==0.115.0
uvicorn==0.30.1
streamlit==1.38.0
faiss-cpu==1.7.4
beautifulsoup4==4.12.3

# Data processing
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1

# Machine learning / embeddings
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.44.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3
